Moved io files into a well-named folder and subfolder (organization).
