+++ src/main/java/io/windows/ErrorWindow.java
@@ -1,4 +1,4 @@
-package CoreEntities.IO;
+package io.windows;
 
 import javax.swing.*;
